--- a/src/generators.jl
+++ b/src/generators.jl
@@ -61,12 +61,12 @@ end
 #offdiag(hssA::HssNode, ::Val{:upper}) = generators(hssA.A11)[1]*hssA.B12*generators(hssA.A11)[2]'
 
 ## orthogonalize generators
-function orthonormalize_generators!(hssA::HssMatrix{T}) where T
+function orthonormalize_generators!(hssA::HssMatrix{T}; multithreaded = HssOptions().multithreaded) where T
   if isleaf(hssA)
     U1 = pqrfact!(hssA.A11.U, sketch=:none); hssA.U = Matrix(U1.Q)
     V1 = pqrfact!(hssA.A11.V, sketch=:none); hssA.V = Matrix(V1.Q)
   else
-    task = RecursionTools.spawn(_orthogonalise_A11!, (hssA,), true)
+    task = RecursionTools.spawn(_orthogonalise_A11!, (hssA,multithreaded), multithreaded)
     if isleaf(hssA.A22)
       U2 = pqrfact!(hssA.A22.U, sketch=:none); hssA.A22.U = Matrix(U2.Q)
       V2 = pqrfact!(hssA.A22.V, sketch=:none); hssA.A22.V = Matrix(V2.Q)
@@ -98,14 +98,14 @@ function orthonormalize_generators!(hssA::HssMatrix{T}) where T
   return hssA
 end
 
-function _orthogonalise_A11!(hssA)
+function _orthogonalise_A11!(hssA, multithreaded)
   if isleaf(hssA.A11)
     U1 = pqrfact!(hssA.A11.U, sketch=:none)
     hssA.A11.U = Matrix(U1.Q)
     V1 = pqrfact!(hssA.A11.V, sketch=:none)
     hssA.A11.V = Matrix(V1.Q)
   else
-    hssA.A11 = orthonormalize_generators!(hssA.A11)
+    hssA.A11 = orthonormalize_generators!(hssA.A11, multithreaded = multithreaded)
     U1 = pqrfact!([hssA.A11.R1; hssA.A11.R2], sketch=:none)
     V1 = pqrfact!([hssA.A11.W1; hssA.A11.W2], sketch=:none)
     rm1 = size(hssA.A11.R1, 1)