@bonevbs 2025-02-06
Great work Baptiste! Do you have any intuition why matvec isn't faster?