@bonevbs 2025-02-06
also - I would suggest bumping up the version number and pushing a new package once this is merged