@BaptisteLamic 2025-02-07
Hi, 
I think that I have been quite conservative in the parallelization, and so far I did not notice any issue when using it for my simulation. So I think that should be fine. The only issue I encountered is the old problem where the code crashes when a block completely vanishes. I may have a look at this over the next few weeks.

Regarding the absence of performance gain for matvec, I suspect that the serial version is simply so fast that the multithreading overhead isn’t worth it. The situation may change for larger problem sizes. Overall, the efficiency of the parallel version is not great, but it still provides a useful speedup.

I agree that could deserve a version bump.